"Fixed" problem with rotation aux task (desc.)
The Rotation auxiliary task had an assert statement to check that the
images were in the [b C H W] format, which asserted that the shape at
index -3 should be the smallest dimension. This is usually true, but not
for the last batch, which might have a single image!

Super juicy bug.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>