Adding the concept of an "Assumption" (see desc.)
Adding a new "assumptions" folder under settings, which for now will
hold things that apply the same way regardless of in a CL or RL setting.

- Moved the train/test loop of ClassIncrementalSetting over to a new
  "Assumption"-ish class, called IncrementalSetting for now.

Still have to add this "assuption" to the RL side of the tree, and then
we shoud be closer to having both branches work the same way!

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>