Added a new 'Sparse' gym Space (see description)
Added a new `Sparse` gym.Space subclass in `sparse_space.py`.

This space is used to wrap other spaces, and when sampled, the returned
value will be set to `None` `none_prob` percentage of the time.

It also alters membership so that `None in Sparse(...)` will always
return `True`, regardless of the type of original space.

When the `sparse_space.py` module is imported, we also modify some of
the utility functions from gym.spaces.utils and gym.vector.utils
"in-place" by making them single-dispatch callables and registering new
implementations of these functions that are only used when passed a
Sparse space object.

For a more visual illustration of this, see the test file, called
`sparse_space_test.py`.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>