Add a --benchmark option as a shortcut (see desc.)
This adds the --benchmark option to Experiment, which is used by
main.py. When passed, the setting's fields are populated with the
values found in the given json or yaml file.

NOTE: If any command-line arguments for the setting are also passed,
an error is raised, as those would have been ignored anyway.

This is really useless for supervised settings, as its basically just
equivalent to setting the --dataset option. It's useful for RL however,
since there's a preset task schedule to be used.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>